Add withDefault to MorphOne relations.
