+++ src/Illuminate/Database/Eloquent/Relations/HasOne.php
@@ -4,17 +4,11 @@ namespace Illuminate\Database\Eloquent\Relations;
 
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Collection;
+use Illuminate\Database\Eloquent\Relations\Concerns\HasDefault;
 
 class HasOne extends HasOneOrMany
 {
-    /**
-     * Indicates if a default model instance should be used.
-     *
-     * Alternatively, may be a Closure or array.
-     *
-     * @var \Closure|array|bool
-     */
-    protected $withDefault;
+    use HasDefault;
 
     /**
      * Get the results of the relationship.
@@ -43,33 +37,6 @@ class HasOne extends HasOneOrMany
     }
 
     /**
-     * Get the default value for this relation.
-     *
-     * @param  \Illuminate\Database\Eloquent\Model  $model
-     * @return \Illuminate\Database\Eloquent\Model|null
-     */
-    protected function getDefaultFor(Model $model)
-    {
-        if (! $this->withDefault) {
-            return;
-        }
-
-        $instance = $this->related->newInstance()->setAttribute(
-            $this->getForeignKeyName(), $model->getAttribute($this->localKey)
-        );
-
-        if (is_callable($this->withDefault)) {
-            return call_user_func($this->withDefault, $instance) ?: $instance;
-        }
-
-        if (is_array($this->withDefault)) {
-            $instance->forceFill($this->withDefault);
-        }
-
-        return $instance;
-    }
-
-    /**
      * Match the eagerly loaded results to their parents.
      *
      * @param  array  $models
@@ -83,15 +50,14 @@ class HasOne extends HasOneOrMany
     }
 
     /**
-     * Return a new model instance in case the relationship does not exist.
+     * Make a new related instance for the given model.
      *
-     * @param  \Closure|array|bool  $callback
-     * @return $this
+     * @param  \Illuminate\Database\Eloquent\Model  $parent
+     * @return \Illuminate\Database\Eloquent\Model
      */
-    public function withDefault($callback = true)
+    public function newRelatedInstanceFor(Model $parent)
     {
-        $this->withDefault = $callback;
-
-        return $this;
+        return $this->related->newInstance()
+            ->setAttribute($this->getForeignKeyName(), $parent->{$this->localKey});
     }
 }
