+++ src/Illuminate/Database/Eloquent/Relations/MorphOne.php
@@ -2,10 +2,14 @@
 
 namespace Illuminate\Database\Eloquent\Relations;
 
+use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Collection;
+use Illuminate\Database\Eloquent\Relations\Concerns\HasDefault;
 
 class MorphOne extends MorphOneOrMany
 {
+    use HasDefault;
+
     /**
      * Get the results of the relationship.
      *
@@ -13,7 +17,7 @@ class MorphOne extends MorphOneOrMany
      */
     public function getResults()
     {
-        return $this->query->first();
+        return $this->query->first() ?: $this->getDefaultFor($this->parent);
     }
 
     /**
@@ -26,7 +30,7 @@ class MorphOne extends MorphOneOrMany
     public function initRelation(array $models, $relation)
     {
         foreach ($models as $model) {
-            $model->setRelation($relation, null);
+            $model->setRelation($relation, $this->getDefaultFor($model));
         }
 
         return $models;
@@ -43,5 +47,18 @@ class MorphOne extends MorphOneOrMany
     public function match(array $models, Collection $results, $relation)
     {
         return $this->matchOne($models, $results, $relation);
+    }
+
+    /**
+     * Make a new related instance for the given model.
+     *
+     * @param  \Illuminate\Database\Eloquent\Model  $parent
+     * @return \Illuminate\Database\Eloquent\Model
+     */
+    public function newRelatedInstanceFor(Model $parent)
+    {
+        return $this->related->newInstance()
+            ->setAttribute($this->getForeignKeyName(), $parent->{$this->localKey})
+            ->setAttribute($this->getMorphType(), $this->morphClass);
     }
 }
